made sub components and sperated the form creation and dropdown logic
